requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
llama-index-core==0.14.14
llama-index-llms-litellm==0.6.*
llama-index-vector-stores-qdrant==0.9.1
llama-index-embeddings-openai==0.5.*
llama-index-readers-file==0.5.6
qdrant-client==1.16.2
redis==5.*
litellm==1.81.*
structlog==25.*
pydantic-settings==2.*
python-dotenv==1.*
pyyaml==6.*
python-multipart==0.0.12
fastembed==0.3.6
pymupdf==1.24.*
sentence-transformers